[feature] Sync finish

+++ sync.py
@@ -1,31 +1,22 @@
-import subprocess
-import os
+#!/usr/bin/env python
 import settings
 import shutil
 import sys
+import inspect
+import history
+import os
 from history import *
+from tools import *
 
 
-def LogError(info, *args, **kwarg):
-    print('[ERROR]')
-    print(info, *args, **kwarg)
-
-
-def Log(*args, **kwarg):
-    print(*args, **kwarg)
-
-
-def sys_call(cmd, showcmd=False, printScreen=True):
-    if showcmd:
-        Log('==>', cmd)
-    output = subprocess.check_output(cmd, shell=True).decode("utf-8").strip(' \n')
-    Log(output)
-    return output
+def git_check_repo_bare():
+    info = sys_call('git status', showcmd=False, printScreen=False)
+    return 'No commits yet' in info
 
 
 def git_check_repo_clean():
     info = sys_call('git status', showcmd=False, printScreen=False)
-    return 'working tree clean' in info
+    return 'nothing to commit' in info
 
 
 def git_check_branch_master():
@@ -35,12 +26,13 @@ def git_check_branch_master():
 
 def git_check_repo_exists():
     info = sys_call('git remote -v | grep origin', showcmd=False, printScreen=False)
-    return info.split('\n')[0].split(' ')[1] == settings.HISTORY_SAVE_REPO
+    return info.split('\n')[0].split(' ')[0].split('\t')[1] == settings.HISTORY_REPO
 
 
 def git_check():
     if not git_check_repo_exists():
-        raise Exception('repo %s not found in directory: %s' % (settings.HISTORY_SAVE_REPO, settings.HISTORY_SAVE_DIR))
+        msg = 'repo %s not found in directory: %s' % (settings.HISTORY_REPO, settings.HISTORY_DIR)
+        raise Exception(msg)
 
     if not git_check_branch_master():
         raise Exception('not on branch master')
@@ -49,45 +41,69 @@ def git_check():
         raise Exception('git directory dirty')
 
 
-def get_path(path):
-    return os.path.abspath(os.path.expanduser(path))
-
-
 def get_time_now_str():
     return datetime.now().strftime('%Y-%m-%d-%S')
 
 
-def sync():
-    shells = sys.argv[1:] if len(sys.argv) > 1 else settings.SHELLS
+def get_shells():
+    shells = []
+    for name in dir(history):
+        cls = getattr(history, name)
+        if 'history' in name.lower() and inspect.isclass(cls) and isinstance(cls(), History):
+            shells.append(name.lower().split('history')[0])
 
-    git_dir = get_path(settings.HISTORY_SAVE_DIR)
+    return [s for s in shells if s]
+
+
+def sync():
+    shells = sys.argv[1:] if len(sys.argv) > 1 else get_shells()
+    Log('shells: ', shells)
+
+    git_dir = full_path(settings.HISTORY_DIR)
     os.chdir(git_dir)
     git_check()
 
-    Log('[ pull history from origin ]')
+    Log('==> [ pull history from origin ]')
     sys_call('git fetch origin')
-    sys_call('git reset --hard origin/master')
+    if not git_check_repo_bare():
+        sys_call('git reset --hard origin/master')
 
-    os.makedirs(get_path(settings.BACKUP_DIR), exist_ok=True)
-    for shell in settings.SHELLS:
-        ShellHistory = globals()[shell.capitalize() + 'History']
-        shell_his_dir = getattr(settings, shell.upper() + '_HISTORY_DIR')
-        shutil.copy(shell_his_dir,
-                    os.path.join(settings.BACKUP_DIR, shell + '_history.' + get_time_now_str()))
-        History.merge_file(shell_his_dir, settings.HISTORY_FILE_NAME, settings.HISTORY_FILE_NAME, cls1=ShellHistory,
-                           cls2=History)
+    Log('==> [ merge histories ]')
+    try:
+        os.makedirs(full_path(settings.BACKUP_DIR), exist_ok=True)
+        if not os.path.exists(settings.HISTORY_FILE_NAME):
+            open(settings.HISTORY_FILE_NAME, 'a').close()
 
-    for shell in settings.SHELLS:
-        ShellHistory = globals()[shell.capitalize() + 'History']
-        shell_his_dir = getattr(settings, shell.upper() + '_HISTORY_DIR')
-        ShellHistory.convert_file(settings.HISTORY_FILE_NAME, shell_his_dir, cls1=History)
+        for shell in shells:
+            Log('[ merge %s... ]' % shell)
+            ShellHistory = globals()[shell.capitalize() + 'History']
+            shell_his_dir = ShellHistory._get_history_dir()
+            shutil.copy(shell_his_dir,
+                        os.path.join(settings.BACKUP_DIR, shell + '_history.' + get_time_now_str()))
+            History.merge_file(shell_his_dir, settings.HISTORY_FILE_NAME, settings.HISTORY_FILE_NAME, cls1=ShellHistory,
+                               cls2=History)
 
-    Log('[ push history to origin ]')
-    sys_call('git add %s' % settings.HISTORY_FILE_NAME)
+        for shell in shells:
+            ShellHistory = globals()[shell.capitalize() + 'History']
+            shell_his_dir = ShellHistory._get_history_dir()
+            ShellHistory.convert_file(settings.HISTORY_FILE_NAME, shell_his_dir, cls1=History)
+    except Exception as e:
+        LogError('[Error while syncing]')
+        sys_call('git clean -d -f')
+        sys_call('git reset --hard master')
+        raise e
+
+    Log('==> [ push history to origin ]')
+    # sys_call('git add %s' % settings.HISTORY_FILE_NAME)
+    sys_call('git add --all')
     if not git_check_repo_clean():
         sys_call('git commit -m "%s save history: %s"' % (get_time_now_str(), shells))
         sys_call('git push origin master')
 
 
 if __name__ == '__main__':
+    # get_shells()
+    # print(get_shells())
     sync()
+    # h = ZshHistory()
+    # print(h.__class__.__name__)
